Move Node to models.
